doc: Organisms 컴포넌트 스토리 작성
- Modal 컴포넌트 스토리 수정
